+++ sdk/core/core-tracing/lib/tracers/noop/noOpBinaryFormat.ts
