+++ sdk/core/core-tracing/lib/tracers/noop/noOpHttpTextFormat.ts
